Undid Changes or minor white lines removal and other insignificant stuff

+++ systems/Collider.java
@@ -8,19 +8,19 @@ public class Collider {
     private int offX;
     private int offY;
 
-    public Collider( int x, int y, int widht, int height, int offX , int offY ){
-        bounds = new Rectangle( x + offX, y + offY, widht , height );
+    public Collider( float x, float y, int widht, int height, int offX , int offY ){
+        bounds = new Rectangle( (int)x + offX, (int)y + offY, widht , height );
         this.widht = widht;
         this.height = height;
         this.offX = offX;
         this.offY = offY;
     }
 
-    public void updateBound( int x, int y){
+    public void updateBound( float x, float y){
         //actualizamos la posicion del cuadro en base al sprite
         //para que cambie cual es el bound actual sobre
         //el cual se puede colisionar
-        bounds.setBounds( x + offX , y + offY, widht, height);
+        bounds.setBounds( (int)x + offX , (int)y + offY, widht, height);
     }
 
     public Rectangle getBounds(){ return bounds; }
